# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Databanka API Documentation",
        "description": "API documentation for project databanka",
        "contact": {
            "email": "fiit.tim09@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/activities/detail/{id}": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Show activity detail",
                "description": "Returns 'activity', 'users', 'registered' and title",
                "operationId": "detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Get event",
                "description": "Returns 'event' by id",
                "operationId": "getEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            },
            "put": {
                "tags": [
                    "Event"
                ],
                "summary": "Update event",
                "description": "Updates event",
                "operationId": "putEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request body does not need to contain all event fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string"
                                    },
                                    "time_to_explain": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        },
        "/events": {
            "put": {
                "tags": [
                    "Event"
                ],
                "summary": "Update events",
                "description": "Updates events",
                "operationId": "putEvents",
                "requestBody": {
                    "description": "Array of JSON object with fields that are going to be updated.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        },
        "/events/{id}/options": {
            "get": {
                "tags": [
                    "EventOptions"
                ],
                "summary": "Get all options of event",
                "description": "Returns 'options' by event id",
                "operationId": "getEventOptions ",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        },
        "/events/detail/{id}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Show event",
                "description": "Returns 'event' with 'helps' associated with them",
                "operationId": "showEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}